Adding documentation and refactoring
