+++ Ising/src/boundary.cpp
@@ -1,13 +1,5 @@
 #include "boundary.hpp"
 
-/**
- * @brief Takes care of the indexing at the boundary of the lattice.
- * 
- * @param ui The row index we "want" to see.
- * @param uj The column index we "want" to see.
- * @param ai The row index the lattice site sees.
- * @param aj The column index the lattice site sees.
- */
 void imposeBC(int w, int h, int ui, int uj, int* ai, int* aj, BoundaryCondition boundary) {
 	switch (boundary) {
 		case BoundaryCondition::PERIODIC:
